% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraintclosure.R
\docType{package}
\name{constraintclosure-package}
\alias{constraintclosure-package}
\alias{constraintclosure}
\title{The 'constraintclosure' package.}
\description{
Research on ecological constraints and community dynamics
}
